requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
neo4j>=5.0.0
pydantic>=2.0.0
pyyaml>=6.0
rapidfuzz>=3.0.0
streamlit>=1.25.0
requests>=2.31.0
transformers>=4.30.0
torch>=2.0.0
pyvis>=0.3.2
networkx>=3.0
